// server/src/types/pokemon.ts
export type Ability = {
    ability: {
      name: string;
      url: string;
    };
    is_hidden: boolean;
    slot: number;
  };
  
  export type Form = {
    name: string;
    url: string;
  };
  
  export type GameIndice = {
    game_index: number;
    version: {
      name: string;
      url: string;
    };
  };
  
  export type Move = {
    move: {
      name: string;
      url: string;
    };
  };
  
  export type Stat = {
    base_stat: number;
    effort: number;
    stat: {
      name:
        | 'hp'
        | 'attack'
        | 'defense'
        | 'special-attack'
        | 'special-defense'
        | 'speed';
      url: string;
    };
  };
  
  export type Type = {
    slot: number;
    type: {
      name: string;
      url: string;
    };
  };
  
  export type Pokemon = {
    abilities: Ability[];
    base_experience: number;
    forms: Form[];
    game_indices: GameIndice[];
    height: number;
    id: number;
    is_default: boolean;
    location_area_encounters: string;
    moves: Move[];
    name: string;
    order: number;
    species: {
      name: string;
      url: string;
    };
    sprites: {
      versions: {
        'generation-v': {
          'black-white': {
            'animated': {
              front_default: string;
            };
        };
      };
      };
      front_default: string;
      other: {
        'dream_world': {
          front_default: string;
        };
        'official-artwork': {
          front_default: string;
        };
        'home': {
          front_default: string;
        };
      };
    };
    stats: Stat[];
    types: Type[];
    weight: number;
  };
  
  export default Pokemon;
  